libtracker-sparql: Check for the correct build CWD as subproject
When building Tracker as a subproject, the CWD at build time is still
that of the root project. We want to check this, instead of the
subproject build root.

This makes the remote module correctly detected in the build dir
when building Tracker as a subproject (as it happens in tracker-miners
CI).