propagate an error if ontology file contains incomplete properties
Make the ontology parser propagate an error if it encounters a property that doesn't have range or domain.
The error also is propagated if the range or domain points to a class that doesn't exist.
It solves the problem of adding properties with no range or domain to the database which causes tracker to crash while the ontology is parsed or when the data is queried.