prefix "unsupported nrl:maxCardinality change" error with line and column numbers
Tracker only support the change of nrl:maxCardinality from 1 to N and not the reverse.
The commit moves the code that check if the nrl:maxCardinality change is supported from tracker_data_ontology_process_changes_pre_db() to tracker_data_ontology_load_statement() while the ontology statements still being parsed.
The code is moved to *_load_statement() as it knows the line which contain the statement that change nrl:maxCardinality, so we can prefix the error message with the current line and column numbers.
It also makes sure that the nrl:maxCardinality value is an integer greater than 0 to avoid any problem that can happen because of a weird value assigned to that property.