DomainIndexAddTest: fix problem of using multipe value property as a domain index
DomainIndexAddTest was using test:a_n_cardinality property as a domain index which is not supported in tracker.
Only single value properties can be used as a domain index.
So, it used test:a_int property instead of test:a_n_cardinality as the domain index for class test:C.