propagate an error if ontology file contain parsing errors
It doesn't stop parsing when an error is found. Instead it prints a warning about the error and continue parsing.
After parsing all ontologies and before making any changes to the db, it propagate an error if one or more parsing errors are faced while parsing.
This approach prints all problems in the ontology file at once, so the developer can fix them all before needing to run tracker many times.