docs: Refactor devhelp generation scripts
The intent is to generate devhelp documentation at dist time so it
is included in the tarball. The steps to generate the devhelp
documentation and fix the searchable keywords is the same, so refactor
this to a common script.

Also, trim a bit the devhelp book, since it can contain a lot of
unneeded files (like 27MB of search data, while devhelp does feature
search itself).