libtracker-data: Remove another argument from insert/delete callback
The object_str argument refers to literal values (i.e. those that are
not rdfs:Resource), nowadays we only track changes to rdf:type properties
so only high-level create/delete/update events are emitted.

We don't need to look into individual literals this way, so the argument
is pointless to have.