libtracker-data: Pass GValue as object instead of GBytes
Between the first things we do with those GBytes are converting them
to a GValue, it becomes a bit awkward that in some places we are
converting GValues to GBytes just to interface with this API.

Make it all GValues, and make it the callers responsibility to
convert from the entry format to GValues that can be consumed here.

As a side effect, the tracker-data-update.c machinery no longer
deals with resources as strings, but plain ROWIDs. It is also
the caller responsibility to obtain a URN first, if required.