libtracker-data: Implement anonymous blank nodes
Now that resources are dealt with as ROWIDs despite them being
named or blank nodes, and that blank nodes don't actually get
a generated URN that is stored in the database, we can make this
a switch to either query on URI + bnode name, or just URI.

This flag enables the blank node semantics defined by the SPARQL
spec, where blank nodes cannot be matched by name.

This is exposed as a TrackerDBManager flag, unused at the moment.