libtracker-data: Hint access cost of "triples" virtual table usage
We are not giving this information back to SQLite in the virtual table
xBestIndex method, so it is left to choose the best strategy to access
the "triples" table without further information, which might come out
wrong.

Hint SQLite about the estimated cost of accessing this table with the
suggested index, so that it always makes the best choice (this happened
to be the case at the time of doing
https://gitlab.gnome.org/GNOME/tracker-miners/-/merge_requests/334, but
seemed to regress substantially at some point).

Accessing the "triples" table works best by providing multiple columns
to match of graph/subject/predicate, so the combination of all values
must get the lowest estimated cost. Likewise, having no columns to match
represents the worst estimated cost. These columns are weighted though,
as providing each of those represents a different reduction in the
amount of disk accesses necessary to return the requested data. These
weights are away by orders of magnitude, in order to ensure SQLite is
strongly hinted of the preference order.

Providing this estimated cost, we can trust that SQLite will use the
best choice available. This should significantly improve (back again)
performance when using the "triples" table in some situations.
"tracker3 info <file>" is one of those, and a rather popular one.