tests: Really test TrackerSparqlStatement on all connections
Since we were using the static TestInfo structs to hold the connection,
we were simply overwriting it in a loop when adding the tests. This made
just the last TrackerSparqlConnection (dbus) to be tested.

Separate this bit, and pass new memory to each individual test, so they
are ensured to use the intended connection. This makes the direct
TrackerSparqlStatement explicitly tested as well.