libtracker-sparql: Seal TrackerNamespaceManagers belonging to a connection
This is a slight API break, but one that is a) unused to my knowledge, and
b) prone to introduce further brokenness if misused. Since any namespace
added this way is still not part of the ontology, it could only be used to
produce RDF that cannot be consumed later on.

Also, there's the implicit assumption that TrackerSparqlConnection is
multithread-aware, but this object would potentially break if it ends up
used and modified from different threads.

So, make tracker_namespace_manager_add_prefix() reserved for namespaces
created through tracker_namespace_manager_new(), and make those coming from
tracker_sparql_connection_get_namespace_manager() immutable.