libtracker-sparql: Deprecate tracker_namespace_manager_get_default()
Since we have tracker_sparql_connection_get_namespace_manager() and allow
different ontologies in TrackerSparqlConnection, this is now the recommended
way to deal with namespaces, even if Nepomuk ones.

TrackerResource still needs to use this namespace to honor the default
behavior when passing a NULL namespaces object in tracker_resource_print_*(),
so reduce the required version there to avoid deprecation warnings when
building Tracker itself.

Fixes: https://gitlab.gnome.org/GNOME/tracker/-/issues/271