libtracker-data: Hint best tracker_triples access strategy for DESCRIBE
With recent SQLite, there's been changes that affected the "best" index
picked when accessing the tracker_triples table, giving up filtering
on the "subject" column when it actually could do that.

Commit 4a08ea0f9b tried to address this making the tracker_triples table
provide an estimation of the cost to access the table with the different
constraints, which fixed some cases like "tracker info <uri>" being able to
filter by the given subject.

But this in turn broke other situations where the tracker_triples table
is used to match with a large number of items, it turns out that SQLite
query planner doesn't distinguish between:

  SELECT FROM triples_table WHERE subject IN ($small_set)

and:

  SELECT FROM triples_table WHERE subject IN ($large_set)

To have a point where using the constraint by "subject" does not make
sense anymore, and it becomes faster to not access the triples_table
by providing many "subject" matches, but query everything altogether.

This is most noticeable on DESCRIBE queries due to the way the
tracker_triples table is joined, which makes them slow on large
datasets.

Since we can't tweak the tracker_triples xBestIndex method to pick
the best behavior here, and the tracker_triples behavior is desirable
on most other places, opt for tweaking the SQLite query planning in the
DESCRIBE query string, so the subject is made ineligible for lookups
in the tracker_triples table there.

This brings the "best" of both worlds for DESCRIBE queries, so the
fastest access method is used depending on whether it is asked to
describe a fixed set of resources or a generic query.