libtracker-sparql: Ensure to register a single instance of serializers
For now, the private TrackerSerializer API was only used in the
TrackerEndpointHttp object, which is compiled as a loadable module since
we support linking with soup 2.x and 3.x simultaneously.

Now that we are introducing usage of this private API from other places
in code, we risk getting the dreaded "type already registered" errors
from GType, since each of libtracker-sparql and the soup modules will
try to register their own.

Make creation of serializers more lenient wrt the GType lookup, so
there's no double registration of the same type, and ensure these GTypes
are initialized from libtracker-sparql so there's no confusion about
who comes first.

This is yet another braindead hack about libsoup loadable modules that
can be removed when we can pick one of them at build time.