libtracker-data: Handle SQLITE_LOCKED_SHAREDCACHE in the update thread
Handling this while executing update statements turns out necessary in the
case of in-memory databases, as it seems the (single) writing thread can
get this SQLITE_LOCKED extended error if there are already existing read
locks. This can be seen (or not, it's quite racy) running
"tracker-miner-fs-3 --dry-run" so it uses an in-memory database.

If this condition happens, we simply retry until we get something else
than SQLITE_LOCKED (i.e. all readers finished). Handling of
this error condition could get fancier (e.g. via sqlite3_unlock_notify()),
but this is already enough to avoid the ill effects.

In order to ensure that this only applies to in-memory connections, check
that we specifically get SQLITE_LOCKED_SHAREDCACHE, since we only use shared
caches for in-memory databases. Getting SQLITE_LOCKED in any other
circumstance is still unexpected and is an error situation, most notably
WAL databases in disk are not affected by this situation.