libtracker-data: Honor out variable in early return path
When handling the insertion of an individual TrackerResource property that
points to another TrackerResource, we check if the latter resource is
already inserted/known and return early in update_resource_single(), but
we forget to set the "id" out variable, thus go with an uninitialized
value here.

Shuffle the early return, so the resource ID is already known, and correctly
returned.