libtracker-data: Shuffle handling of properties not belonging to a class
We currently warn if we are setting a property on a resource that is not
expected due to missing rdf:type. This is only relevant while handling
insertion of new data, all other places use get_old_property_values() in
places where either data is being deleted, or we know we deal with
properties in the domain of the resource.

Move this error to the place where it should be handled, instead of doing
it generically from get_old_property_values().