libtracker-data: Forward annotated column types on SELECT queries
Forward this additional information in the query itself, as extra columns
after the user visible ones. We already have ":type" suffixed subvariables
for the places where we cannot infer that from the query being parsed,
prefer these to the TrackerPropertyType found out at parse time, as that
will contain the real type for the given column at the current row.