libtracker-sparql: Ensure correct integer ordering of queried notifier IDs
We query info for IDs in an specific order and expect them to come in the exact
same order in the cursor, however the column may be implicitly mapped to a
string type, which may result in a different ordering.

Ensure the ORDER BY happens with integer affinity, this ensures the cursor
values are in the expected order.

Fixes: https://gitlab.gnome.org/GNOME/tracker/-/issues/345
Related: https://gitlab.gnome.org/GNOME/gnome-music/-/issues/487