build: Add an option to control building tests
Commit 8569c99d2926 ("Look for a Python installation with the tap
module") correctly added a dependency on the gi Python module for tests.
Commit 51ca035697a4 ("Make tracker:endpoint test work on macOS")
restructured Python probing by moving it to the top-level meson.build.

These two commits have the side-effect of exposing the fact that the
gi Python module is really only a test dependency, and should not be
required if building without tests.

As such, add a -Dtests=bool option to disable building tests. This
reduces the number of ninja targets from 232 to 188 on my non-test
build.