ci: Add repo-sanity check
It is somewhat easy to dismiss failed build images as errors in
the pipeline rather than errors in the repository configuration.

Make a separate check in the review stage testing that the user
repository is in a state able to run the rest of CI (e.g. generating
or storing images in its own container registry, since this is
necessary with ci-templates).

The error message is made known through both junit and CI job output.