ci: Enable detached merge request pipelines
The check jobs are relying on environment variables that are only
set on merge request pipelines, so they've been kinda ineffective
so far.

Enable the detached merge request pipelines as documented at
https://docs.gitlab.com/ee/ci/pipelines/merge_request_pipelines.html
so that these work. Since we want all jobs to run in merge pipelines,
using "workflow" is good for us.