ci: Update alpine image, and use "latest"
The latter seems necessary for ci-templates to pan along (and anyways,
sticking to "edge" for things that we want reproduceability and stability
like CI is not recommended).

This went unnoticed since this image hasn't been rebuilt for some
time, but has probably been broken since the last update to ci-templates
dependencies, and can be seen in merge requests from user repositories.