libtracker-data: Pass correct variable name for non-direct types
Property types like dates or resources use wrapper functions to print
the user visible values, so the SQL query ends up as:

  SELECT SparqlFormatTime("v_date") ...

And results in those variables having the unintended name (e.g.
`SparqlFormatTime("v_date")` instead of `date`. Ensure these (and
any other variable, just redundantly) get an AS alias so we get
the right variable name propagated. E.g.:

  SELECT SparqlFormatTime("v_date") AS "date" ...

It is worth noting that in other places, the internal variable
name (e.g. `v_date`) is passed to the upper layers, and the v_
prefix is removed in TrackerDBCursor code. This cannot be done
here esp. for date types, since SQL places GROUP/ORDER BY at a
different level than SPARQL, so these manage to pick up the
user-visible strings instead of the more accurate timestamps.
This is not intended, so produce the final variable name in
place so the SQL does not get confused about that variable to
use.

Fixes: https://gitlab.gnome.org/GNOME/tracker/-/issues/352