ci: Run coverage on all test jobs
And have the coverage job merge all results. So far this is mostly
repeated work with maybe some slight timing-related variances in
code paths, but this may be useful to get an unified coverage report
for different compilation flags in the future.

So far, Ubuntu seems to have an old gcovr version that generates
a different JSON intermediate file. Since gcovr requires that all
JSON versions match, Ubuntu is so far kept out.