ci: Work around rawhide failures by using crun runners
With a combination of recent glib, and old runc/seccomp in the
runners, we run into the error messages described at:
https://gitlab.gnome.org/Infrastructure/GitLab/-/issues/545

So far, the advised workaround is to use the `crun` tag to use
runners that don't run that combination of software. Add this
specifically for the rawhide test job.