libtracker-sparql: Fix libtracker-remote-soup{2,3} modules linking
The modules should link dynamically and not statically with
libteacker-sparql. Linking statically results in the modules and the
shared library to export definitions for the same GObjects, resulting
in type registration errors.

The tracker-remote modules however need access to the private
TrackerSerializer classes. Unless these GObject are made public, their
definition must be linked in the modules too. As the code implementing
TrackerSerializer and the specialized version is pretty small, the
simplest solution is to compile the sources twice, without going
through an intermediate library. The definitions being private to
libtracker-sparql and to the libtracker-remote-soup{2,3} modules,
having two defintions for them does not cause issues.

Fixes #358 and likely #350 too.