build: Generalize hack to ensure soup modules are a dependency of GIR
We want to avoid meson version bumps, but this hack to ensure that GIR
generation depends on the soup modules (since they will be pulled at
runtime by g-ir-scanner during introspection) does not work across all
meson versions.

Use different hacks dependent on the meson version, so that the modules
are a build-time dependency of the GIR target.