libtracker-fts: Drop in-tree fts5 module
This was only a last resort for sqlite < 3.20, since module loading
API changed around that version, also it probably did us a disservice
since we barely updated the fts5.c file.

We no longer accept such old versions and we expect that sqlite has
a builtin FTS5 module, so we can drop this.