build: Depend on sqlite >=3.20.0
Back in the day, 3.20.0 was stablished as the baseline to expect
a builtin fts5 module, due to module loading API changes in sqlite.
Fast forward to today, this version was released in Aug 2017 so it
is old enough to bump the dependency. This will allow us to drop our
in-tree fts5 module.