libtracker-sparql: Build as a monolithic library
We want to increase the interactions between different parts of
the code tree (e.g. unify serialization/deserialization infrastructure
so it's usable by everyone) without neccessarily making the different
pieces of the mechanism public API.

By building all subcomponents as static libraries, we were constrained
to the API that we wanted to have publicly available.

Turn libtracker-sparql into a library that is built monolithically
in order to lift that barrier within internals.

We anyhow still need access to private API in the tracker3 sql/export
utilities, make this still through a static library, but one that is
right underneath libtracker-sparql, instead of being a selected cut.

And sadly, we cannot get rid yet of the static libtracker-data library
for the http endpoint modules (a requirement via its TrackerSparqlStatement
implementation). Keep this static library solely for this.