libtracker-sparql: Add base http helper object definitions
This is a small internal API consisting of a TrackerHttpClient and
a TrackerHttpServer abstract objects. These are meant to wrap the
interaction with libsoup in a way that is most isolated from internal
library workings.

These objects will have final implementations living in modules, but
also will be interacted through the base API from the library code,
this so far adds the latter.