tests: Do not use separate test to poke cursors after last next()
Currently, the behavior of tracker_sparql_cursor_get*() over an already
finished cursor is inconsistent. All return NULL, but some trigger a
warning and some don't.

Since we want to run these tests on all connections, at the moment
drop this test that asserted for failures/warnings. We might want to
find a consistent behavior and test for it, and if it turns out to
involve runtime warnings, it is best to use g_test_expect_message()
than subprocesses.