libtracker-data: Ensure blank node row ids are integer typed
Since commit 6cf3168302350f3fd3ef7f8bf79022950563300e tracker sometimes
returns an unqualified row id to applications that are asking for
a fully qualfied blank node urn.

This is because sometimes it generates URN with a text type instead of
an integer type, and so `SparqlPrintIRI` fails to recognize it as a
blank node identifier.

The result is that gnome-photos fails to match up album collections internally
between different queries, because sometimes tracker returns the fully
qualified urn and sometimes it returns the bare row id.

This commit address the problem by introducing a cast in the query that is
returning the row id as text instead of as an integer.

Fix suggested by Carlos Garnacho.

Closes https://gitlab.gnome.org/GNOME/tracker/-/issues/363