tests: Add tests for queries around blank nodes
We have 2 behaviors to test here, the default non-standard behavior
that treats blank nodes as the equivalent of a lazy IRI generator,
and the spec-compliant behavior that was made possible through the
TRACKER_SPARQL_CONNECTION_FLAGS_ANONYMOUS_BNODES in 3.3.x.

In the former, we expect the engine to treat blank nodes like IRIs
when passed as literals, e.g.:

  SELECT (<urn:bnode:...> AS ?u) { }
  SELECT ?p ?o { <urn:bnode:...> ?p ?o }

should identify the correct resource from the blank node name
and return data around it, making it effectively identifiable.

In the latter, blank nodes are tightly scoped to the result set,
a blank node existing in a result set just means something exists
that matches the graph patter, but it cannot be kept and poked
around in future queries like the above.

Add tests to ensure that both behaviors are gotten right.