libtracker-sparql: Refactor handling of prefix/base in Turtle parsing
This is handled in the same state that handles subject, but since we
handle Trig format in the same object, this should happen prior to
parsing the first GRAPH clause there.

Move this parsing to the initial state, so that it applies for both
formats the same. This involved changes in parser location handling
for errors, since the initial state now actually advances the parser.