core: Handle conversion of values for superproperties with differing types
This is not quite correct, but it is the most straightforward way to sort
of handle situations in our own Nepomuk ontology that superproperties have
different rdfs:range than their subproperties.

For these cases, handle the conversion of the value. Since the worst case
(and the one we happen to exercise) is about converting resource ROWIDs to
strings, handle this specific situation.

This helps in serialization and deserialization of this data, since the
ROWIDs were not something that could be replicated in other databases
without mismatches (e.g. the resource having a different ROWID in the
new database importing the data) or worse (e.g. cardinality or other
constraints broken).