libtracker-data: Fetch nrl:modified from all graphs
We are just looking at the default graph when initializing nrl:modified
after opening an existing database. Since the default graph possibly
only contains part of the data, we are possibly/likely missing the latest
sequence number if used on other graphs.

Check all graphs here relying on our virtual tracker_triples table, so
we are ensured to get the actual latest/greatest nrl:modified sequence
that was previously used from the union of all graphs. This ensures us
to get always a fresh sequence number for newly inserted data.