utils: Drop tracker-resdump noinst util
This small utility prints Turtle for resources in the miner fs
database. It's from the olden days that introspection through
SPARQL queries was limited, we didn't have RDF production
capabilities, and there was an unique store.

We currently have this feature built in, and available through
"tracker3 export", there is no need for this ad-hoc util executable
anymore.