libtracker-sparql/bus: Initialize string length out value in all paths
Fill in the out parameter in tracker_sparql_cursor_get_string() also
on NULL strings. We used to do that, and even had code that internally
relied on that.

Fixes accessing through the portal after the bus connection rewrite,
since the TrackerEndpointDBus in the portal would rely on prior
behavior to correctly write the cursor to the sandboxed client.

Sadly, our puny portal tests didn't catch this situation.