libtracker-sparql/core: Loosen TrackerDBStatement ownership restrictions
Currently, we enforce that every TrackerDBStatement is either executed,
reset, or fully iterated before considering the statement ready for use
elsewhere again. We enforce this even with TrackerDBStatements that are
requested not to be cached.

Since we have a number of relatively invariable internal queries and
updates (e.g. resource refcount maintenance), it could be benefitial for
those to have the update machinery able to keep TrackerDBStatements
around, instead of building query strings just to be able to get a
statement from the DB interface.

In order to allow for this usecase, soften the restrictions about a
TrackerDBStatement being only able to be executed once for uncached
statements (i.e. TRACKER_DB_STATEMENT_CACHE_TYPE_NONE). Callers will
be able to keep them around for multiple executions.

Since the variable name carrying this accounting had a slightly
misleading name (the cache owns the statements, just borrows them for
execution), also change the name so it is more intuitively correct.