libtracker-sparql/core: Set nrl:modified/added consistently
For nrl:modified, we had some double handling since the modifier was
pushed as a property change, but also dealt with directly when flushing
the buffer.

Since these are essentially properties (even though changed automatically),
it makes sense to stick to the former, also for nrl:added. Do that, and
drop the special handling of these 2 properties when flushing the buffer.

This commit also changes the expected output of a serialization test,
since nrl:added is added now on more graphs than the first one that sees
a resource, this slightly changes the spacing after filtering the test
filters nrl:added/nrl:modified.