core: Refactor buffering of database updates
Currently, our caching of triples does have a number of nested structures:

- In the buffer there is a struct for each graph
- In the graph struct there is a set of changed resource
- In the resource struct there is a set of modified tables
- In the table struct there is a set of modified properties
- In the property struct there is a list of values

This incurs in a maintenance cost that is higher than desired, adding and
removing elements here becomes a fair chunk of the time spent in updates,
since there is a number of allocations and list/hashtable updates
performed for batches that deal with a fair amount of different resources
(i.e. most of them).

In order to improve this, keep use two arrays to buffer this data:

- A "properties" array, that keeps individual predicate/object pairs. This
  is used to store the values of properties being inserted or deleted for
  single-valued and multivalued properties. This struct is "linked" with
  (i.e. references) other elements in the array, so that e.g. class updates
  may reference multiple properties/values being updated.
- An "update log" array, containing structs that are a event_type/graph/
  subject tuple, plus optionally a link to one of the properties in the
  previous array, all other properties are fetched through iterating through
  the linked properties. These log entries are valid for class table updates
  (i.e. single-valued properties) or multi-valued property tables.

These arrays make allocating the buffer a one-time operation (buffer size
is fixed, and the arrays are reused during the processing of a TrackerBatch)
and insertions into the log largely O(1) as opposed to a number of
array/hashtable lookups and inserts.

But we still want to coalesce updates to a same class table (e.g. changes to
several single-valued properties in the same table), for that there is an
additional hashtable set that uses these log entries as keys themselves,
with special hash/equal functions, lookups for prior events modifying the
same TrackerClass is also quite fast.

Overall, this makes the maintenance of this buffer less expensive in the
big picture. Even though there are still some remnants of the previous
caching for graphs and resources, this plays less of a role.

Since this changes the ordering of updates, some tests that rely on implicit
ordering (DESCRIBE ones) had to be adapted for this change.